+++ changes.md
@@ -7,6 +7,8 @@ Change Log
 * added an Option[Logger] to readBundle()
 * pass an implicit Logger to all that need it
 * removed the requirement for .json file extension when processing zip files
+* added a Counter class to count the SDO, SRO and StixObj processed
+
 
 ### changes in 0.1